requirements.txt
agithub==2.1
appdirs==1.4.3
beautifulsoup4==4.6.0
bs4==0.0.1
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
PyYAML==3.12
six==1.10.0
